# perf(events): get_events performance & Google Calendar Integration fixes