--- a/frappe/desk/doctype/event/README.md
+++ b/frappe/desk/doctype/event/README.md
@@ -1,1 +1,0 @@
-Calendar Event