--- a/frappe/desk/doctype/event/event.py
+++ b/frappe/desk/doctype/event/event.py
@@ -398,13 +398,10 @@ def get_events(
 				target_date = add_days(target_date, 1)
 
 		elif e.repeat_on == "Weekly":
-			first_occurence_in_range = e.starts_on.date()
-			jump_ahead = abs((first_occurence_in_range - start).days // 7)
-			target_date = add_days(first_occurence_in_range, 7 * jump_ahead)
-
+			target_date = start
 			while target_date <= end:
 				resolve_event(e, target_date=target_date, repeat_till=repeat_till)
-				target_date = add_days(target_date, 7)
+				target_date = add_days(target_date, 1)  # Increment by 1 to capture multiple days in the week
 
 		elif e.repeat_on == "Monthly":
 			first_occurence_in_range = e.starts_on.date()