… frappe/desk/doctype/event/event.json
… frappe/tests/utils/generators.py