Started as a backport of #31772 with conflicts but I had to rewrite newer parts of the endpoint that only existed on develop. Added more thorough test cases and re-rewrote the original backport after uncovering buggy behaviour in previous version and previous implementations (including the current).

---

### Highlights

* Makes `frappe.desk.doctype.event.event.get_events` 14-15x faster [benchmark holds up]
* URLs under `Event.google_meet_link` are much bigger nowadays - updated fieldtype to Small Text
* Don't sync recurring events from Google Calendars since Frappe doesn't support relevant features - faster syncs & de-duplicated events
* Fix `get_events`'s quirks that seemed like clear bugs
* Add more test cases

```python
# before
In [2]: %timeit frappe.call("frappe.desk.doctype.event.event.get_events", doctype="Event", start="2025-03-31", end="2025-05-12", filters=[["Event","status","=","Open", False]
   ...: ], user='gavin.dsouza@simplesimple.org')
34.9 ms ± 2.34 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# after
In [2]: %timeit frappe.call("frappe.desk.doctype.event.event.get_events", doctype="Event", start=date(2025, 3, 31), end=date(2025, 5, 12), filters=[["Event","status","=","Ope
   ...: n", False]], user='gavin.dsouza@simplesimple.org')
2.41 ms ± 51.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

A breaking change that should be a problem in rare cases - if get_events is being used programmatically and passes `start` & `end` as string, this change breaks support for that🤞🏼 

> NOTE: I re-raised this PR from the backport mergify created since this became more unrelated to the initial backported changes. ref: https://github.com/frappe/frappe/pull/32116